meger from original

+++ src/extension/resource/Resource.ts
@@ -1,20 +1,31 @@
-/**
- * Copyright (c) Egret-Labs.org. Permission is hereby granted, free of charge,
- * to any person obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish, distribute,
- * sublicense, and/or sell copies of the Software, and to permit persons to whom
- * the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
- * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
- * FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
- * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
+//////////////////////////////////////////////////////////////////////////////////////
+//
+//  Copyright (c) 2014-2015, Egret Technology Inc.
+//  All rights reserved.
+//  Redistribution and use in source and binary forms, with or without
+//  modification, are permitted provided that the following conditions are met:
+//
+//     * Redistributions of source code must retain the above copyright
+//       notice, this list of conditions and the following disclaimer.
+//     * Redistributions in binary form must reproduce the above copyright
+//       notice, this list of conditions and the following disclaimer in the
+//       documentation and/or other materials provided with the distribution.
+//     * Neither the name of the Egret nor the
+//       names of its contributors may be used to endorse or promote products
+//       derived from this software without specific prior written permission.
+//
+//  THIS SOFTWARE IS PROVIDED BY EGRET AND CONTRIBUTORS "AS IS" AND ANY EXPRESS
+//  OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+//  OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+//  IN NO EVENT SHALL EGRET AND CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+//  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+//  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;LOSS OF USE, DATA,
+//  OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+//  LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+//  NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
+//  EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+//////////////////////////////////////////////////////////////////////////////////////
 
 
 module RES {
@@ -120,10 +131,11 @@ module RES {
      * 销毁单个资源文件或一组资源的缓存数据,返回是否删除成功。
      * @method RES.destroyRes
      * @param name {string} 配置文件中加载项的name属性或资源组名
+     * @param force {boolean} 销毁一个资源组时其他资源组有同样资源情况资源是否会被删除，默认值true
      * @returns {boolean}
      */
-    export function destroyRes(name:string):boolean{
-        return instance.destroyRes(name);
+    export function destroyRes(name:string, force?:boolean):boolean{
+        return instance.destroyRes(name, force);
     }
     /**
      * 设置最大并发加载线程数量，默认值是2.
@@ -574,11 +586,12 @@ module RES {
          * 销毁单个资源文件或一组资源的缓存数据,返回是否删除成功。
 		 * @method RES.destroyRes
          * @param name {string} 配置文件中加载项的name属性或资源组名
+         * @param force {boolean} 销毁一个资源组时其他资源组有同样资源情况资源是否会被删除，默认值true
 		 * @returns {boolean}
          */
-        public destroyRes(name:string):boolean{
+        public destroyRes(name:string, force:boolean = true):boolean{
             var group:Array<any> = this.resConfig.getRawGroupByName(name);
-            if(group){
+            if(group && group.length > 0){
                 var index:number = this.loadedGroups.indexOf(name);
                 if(index!=-1){
                     this.loadedGroups.splice(index,1);
@@ -586,9 +599,15 @@ module RES {
                 var length:number = group.length;
                 for(var i:number=0;i<length;i++){
                     var item:any = group[i];
-                    item.loaded = false;
-                    var analyzer:AnalyzerBase = this.getAnalyzerByType(item.type);
-                    analyzer.destroyRes(item.name);
+                    if(!force && this.isResInLoadedGroup(item.name)) {
+
+                    }
+                    else {
+                        item.loaded = false;
+                        var analyzer:AnalyzerBase = this.getAnalyzerByType(item.type);
+                        analyzer.destroyRes(item.name);
+                        this.removeLoadedGroupsByItemName(item.name);
+                    }
                 }
                 return true;
             }
@@ -599,8 +618,42 @@ module RES {
                 item = this.resConfig.getRawResourceItem(name);
                 item.loaded = false;
                 analyzer = this.getAnalyzerByType(type);
-                return analyzer.destroyRes(name);
+                var result = analyzer.destroyRes(name);
+                this.removeLoadedGroupsByItemName(item.name);
+                return result;
             }
+        }
+        private removeLoadedGroupsByItemName(name:string):void {
+            var loadedGroups:Array<string> = this.loadedGroups;
+            var loadedGroupLength:number = loadedGroups.length;
+            for(var i:number = 0 ; i < loadedGroupLength ; i++) {
+                var group:Array<any> = this.resConfig.getRawGroupByName(loadedGroups[i]);
+                var length:number = group.length;
+                for(var j:number = 0 ; j < length ; j++) {
+                    var item:any = group[j];
+                    if(item.name == name) {
+                        loadedGroups.splice(i, 1);
+                        i--;
+                        loadedGroupLength = loadedGroups.length;
+                        break;
+                    }
+                }
+            }
+        }
+        private isResInLoadedGroup(name:string):boolean {
+            var loadedGroups:Array<string> = this.loadedGroups;
+            var loadedGroupLength:number = loadedGroups.length;
+            for(var i:number = 0 ; i < loadedGroupLength ; i++) {
+                var group:Array<any> = this.resConfig.getRawGroupByName(loadedGroups[i]);
+                var length:number = group.length;
+                for(var j:number = 0 ; j < length ; j++) {
+                    var item:any = group[j];
+                    if(item.name == name) {
+                        return true;
+                    }
+                }
+            }
+            return false;
         }
         /**
          * 设置最大并发加载线程数量，默认值是2.
